+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -22,8 +22,9 @@ public interface FilmStorage {
 
     Film getFilmById(Long id);
 
-    Collection<Film> getFilmByDirectorIdSortedByYear(Integer id);
+    Collection<Film> getFilmByDirectorIdSortedByYear(Long id);
 
-    Collection<Film> getFilmByDirectorIdSortedByLikes(Integer id);
+    Collection<Film> getFilmByDirectorIdSortedByLikes(Long id);
+
     List<Film> getCommonFilms(Long userId, Long friendId);
 }
